pyproject.toml
[tool.poetry]
name = "crypto-price-predictor"
version = "0.1.0"
description = ""
authors = ["Raphael Tang <raphpb1912@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "1.19.5"
pandas = "1.3.3"
matplotlib = "^3.4.3"
pandas-datareader = "^0.10.0"
sklearn = "^0.0"
tensorflow = "2.2"
seaborn = "^0.11.2"
python-telegram-bot = "^13.7"
python-dotenv = "^0.19.1"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
